phamhai20_11_2024

+++ back-end/payment-service/src/main/java/Du/An/Ban/Sach/Tasc/payment_service/service/VNPayService.java
@@ -132,7 +132,7 @@ public class VNPayService {
         }
     }
 
-    public String refundTransaction(String transactionNo, long amount, String createdBy, String description) {
+    public String refundTransaction(String transactionNo, double amount, String createdBy, String description) {
         String vnp_Command = "refund";
         String vnp_TmnCode = VNPayConfig.vnp_TmnCode;
         String vnp_HashSecret = VNPayConfig.vnp_HashSecret;
